+++ Controls-demo/MasterDetail/ContrastBackground/Index.wml
@@ -1,136 +1,26 @@
-<div class="controlsDemo__wrapper controls-demo_MasterDetail-background">
-   <div class="controlsDemo__masterDetail-container">
-      <div class="controls-text-label">grid:View в мастере</div>
-      <Controls.masterDetail:Base>
-         <ws:master>
-            <Controls.scroll:Container shadowMode="js">
-               <Controls.masterDetail:List bind:markedKey="_markedKey">
-                  <Controls.grid:View
-                          source="{{_masterSource}}"
-                          style="master"
-                          backgroundStyle="master"
-                          keyProperty="id"
-                          markerVisibility="visible"
-                          itemActions="{{_itemActions}}"
-                          columns="{{_gridColumns}}"
-                  >
-                  </Controls.grid:View>
-               </Controls.masterDetail:List>
-            </Controls.scroll:Container>
-         </ws:master>
-         <ws:detail>
-            <Controls.scroll:Container attr:class="controlsDemo_fixedWidth500">
-               <Controls.list:View
-                       keyProperty="id"
-                       source="{{_detailSource}}"
-                       backgroundStyle="detail">
-                  <ws:filter myOpt="{{detail.selectedMasterValue || '0'}}">
-                  </ws:filter>
-                  <ws:itemTemplate>
-                     <ws:partial template="Controls/list:ItemTemplate">
-                        <ws:contentTemplate>
-                           <ws:partial
-                                   item="{{itemTemplate.item.contents}}"
-                                   template="wml!Controls-demo/MasterDetail/itemTemplates/TaskItemTemplate"
-                           />
-                        </ws:contentTemplate>
-                     </ws:partial>
-                  </ws:itemTemplate>
-               </Controls.list:View>
-            </Controls.scroll:Container>
-         </ws:detail>
-      </Controls.masterDetail:Base>
-   </div>
-   <div class="controlsDemo__masterDetail-container">
-      <div class="controls-text-label">list:View в мастере</div>
-      <Controls.masterDetail:Base>
-         <ws:master>
-            <Controls.scroll:Container shadowMode="js">
-               <Controls.masterDetail:List bind:markedKey="_markedKey1">
-                  <Controls.list:View
-                          source="{{_masterSource}}"
-                          style="master"
-                          backgroundStyle="master"
-                          keyProperty="id"
-                          markerVisibility="visible"
-                          itemActions="{{_itemActions}}"
-                          columns="{{_gridColumns}}">
-                     <ws:itemTemplate>
-                        <ws:partial template="Controls/list:ItemTemplate">
-                           <ws:contentTemplate>
-                              <div class="controls-demo_MasterDetail-item-list">{{itemTemplate.item.contents.name}}</div>
-                           </ws:contentTemplate>
-                        </ws:partial>
-                     </ws:itemTemplate>
-                  </Controls.list:View>
-               </Controls.masterDetail:List>
-            </Controls.scroll:Container>
-         </ws:master>
-         <ws:detail>
-            <Controls.scroll:Container attr:class="controlsDemo_fixedWidth500">
-               <Controls.list:View
-                       keyProperty="id"
-                       source="{{_detailSource}}"
-                       backgroundStyle="detail">
-                  <ws:filter myOpt="{{detail.selectedMasterValue || '0'}}">
-                  </ws:filter>
-                  <ws:itemTemplate>
-                     <ws:partial template="Controls/list:ItemTemplate">
-                        <ws:contentTemplate>
-                           <ws:partial
-                                   item="{{itemTemplate.item.contents}}"
-                                   template="wml!Controls-demo/MasterDetail/itemTemplates/TaskItemTemplate"
-                           />
-                        </ws:contentTemplate>
-                     </ws:partial>
-                  </ws:itemTemplate>
-               </Controls.list:View>
-            </Controls.scroll:Container>
-         </ws:detail>
-      </Controls.masterDetail:Base>
-   </div>
-   <div class="controlsDemo__masterDetail-container">
-      <div class="controls-text-label">treeGrid:View в мастере</div>
-      <Controls.masterDetail:Base>
-         <ws:master>
-            <Controls.scroll:Container shadowMode="js">
-               <Controls.masterDetail:List>
-                  <Controls.treeGrid:View
-                          source="{{_masterSource}}"
-                          style="master"
-                          backgroundStyle="master"
-                          keyProperty="id"
-                          parentProperty="Раздел"
-                          nodeProperty="Раздел@"
-                          markerVisibility="visible"
-                          itemActions="{{_itemActions}}"
-                          columns="{{_gridColumns}}"
-                  >
-                  </Controls.treeGrid:View>
-               </Controls.masterDetail:List>
-            </Controls.scroll:Container>
-         </ws:master>
-         <ws:detail>
-            <Controls.scroll:Container attr:class="controlsDemo_fixedWidth500">
-               <Controls.list:View
-                       keyProperty="id"
-                       source="{{_detailSource}}"
-                       backgroundStyle="detail">
-                  <ws:filter myOpt="{{detail.selectedMasterValue || '0'}}">
-                  </ws:filter>
-                  <ws:itemTemplate>
-                     <ws:partial template="Controls/list:ItemTemplate">
-                        <ws:contentTemplate>
-                           <ws:partial
-                                   item="{{itemTemplate.item.contents}}"
-                                   template="wml!Controls-demo/MasterDetail/itemTemplates/TaskItemTemplate"
-                           />
-                        </ws:contentTemplate>
-                     </ws:partial>
-                  </ws:itemTemplate>
-               </Controls.list:View>
-            </Controls.scroll:Container>
-         </ws:detail>
-      </Controls.masterDetail:Base>
-   </div>
+<div class="controlsDemo__wrapper">
+    <div class="controlsDemo__cell">
+        <Router.router:Reference attr:class="controls-text-label"
+                                 state="/Controls-demo/app/:app"
+                                 app="{{ 'Controls-demo/MasterDetail/ContrastBackground/grid/Index' }}">
+            <a href="{{content.href}}">grid:View в мастере</a>
+        </Router.router:Reference>
+        <Controls-demo.MasterDetail.ContrastBackground.grid.Index/>
+    </div>
+    <div class="controlsDemo__cell">
+        <Router.router:Reference attr:class="controls-text-label"
+                                 state="/Controls-demo/app/:app"
+                                 app="{{ 'Controls-demo/MasterDetail/ContrastBackground/list/Index' }}">
+            <a href="{{content.href}}">list:View в мастере</a>
+        </Router.router:Reference>
+        <Controls-demo.MasterDetail.ContrastBackground.list.Index/>
+    </div>
+    <div class="controlsDemo__cell">
+        <Router.router:Reference attr:class="controls-text-label"
+                                 state="/Controls-demo/app/:app"
+                                 app="{{ 'Controls-demo/MasterDetail/ContrastBackground/treeGrid/Index' }}">
+            <a href="{{content.href}}">treeGrid:View в мастере</a>
+        </Router.router:Reference>
+        <Controls-demo.MasterDetail.ContrastBackground.treeGrid.Index/>
+    </div>
 </div>
